requirements.txt
numpy==1.26.1
pandas==2.1.3
matplotlib==3.8.1
seaborn==0.13.0
scikit-learn==1.3.2

fastapi==0.109.0
uvicorn==0.25.0
scikit-learn==1.3.2

ipython==8.17.2
ipykernel==6.26.0
ipywidgets==8.1.1


SQLAlchemy==2.0.23
sqlalchemy_utils==0.41.1

python-multipart
bcrypt==4.0.1
python-jose[cryptography]
social-auth-core[openidconnect]
passlib[bcrypt]
